+++ src/ui/screens/passkeys/mod.rs
@@ -1,3 +1,3 @@
 pub mod view;
 pub mod view_model;
-pub use view_model::{PasskeysEvent, PasskeysView};
+pub use view_model::{PasskeysEvent, PasskeysViewModel};
